refactor: новая структура частей стора
