+++ src/components/burger-constructor/stuffing/stuffing.tsx
@@ -4,7 +4,7 @@ import type { Identifier, XYCoord } from 'dnd-core';
 import { ConstructorElement, DragIcon } from '@ya.praktikum/react-developer-burger-ui-components';
 
 import { TIngredientInOrder } from '../../../types/ingredient';
-import { removeStuffing } from '../../../services/redusers/order';
+import { removeStuffing } from '../../../services/order/order.reducer';
 import { useAppDispatch } from '../../../hooks/use-store';
 
 import styles from './stuffing.module.css';
