+++ src/pages/app/app.tsx
@@ -14,10 +14,10 @@ import Profile from '../profile/profile';
 import PrivateRoute from '../private-route';
 import OnlyNotAuthRoute from '../only-not-auth-route';
 import IngredientDetails from '../../components/ingredient-details/ingredient-details';
-import { fetchGetItems } from '../../services/redusers/app';
+import { fetchGetItems } from '../../services/app/app.reducer';
 import Modal from '../../components/modal/modal';
 import ProfileEditor from '../profile/profile-editor';
-import { getUser } from '../../services/redusers/auth';
+import { getUser } from '../../services/auth/auth.reducer';
 import TLocation from '../../types/location';
 import { useAppDispatch } from '../../hooks/use-store';
 import Feed from '../feed/feed';
