+++ src/pages/profile/orders.tsx
@@ -1,7 +1,7 @@
 import FeedOrders from '../../components/orders-feed/feed-orders';
 import { useAppSelector } from '../../hooks/use-store';
 import { privateFeedWsApi } from '../../services/api/private-live-feed';
-import { ordersSelector } from '../../services/selectors/history';
+import { ordersSelector } from '../../services/order-history/history.selector';
 
 import styles from './profile.module.css';
